시세다운

+++ model/day.go
@@ -7,5 +7,5 @@ type Day struct {
 	LowPrice 		int
 	ClosePrice   	int
     Volume 			int
-	ForeignerBurnoutRate float32
+	ForeignerBurnoutRate float64
 }
